attendai/backend/requirements.txt
# ═══════════════════════════════════════════════════════
#  AttendAI Backend — Python Dependencies
# ═══════════════════════════════════════════════════════

# ─── Web Framework ──────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# ─── Database ───────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# ─── Validation & Settings ──────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ─── Security & Auth ────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.7

# ─── Cache & Queue ──────────────────────────────────────
redis[hiredis]==5.0.4
celery[redis]==5.4.0
kombu==5.3.5

# ─── Face Recognition & CV ──────────────────────────────
insightface==0.7.3
onnxruntime==1.18.0        # CPU; swap to onnxruntime-gpu for GPU
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0
faiss-cpu==1.8.0           # faiss-gpu для GPU-версии

# ─── File Storage (S3/MinIO) ────────────────────────────
boto3==1.34.100
aiobotocore==2.13.0

# ─── Reports ────────────────────────────────────────────
reportlab==4.2.0
openpyxl==3.1.2
jinja2==3.1.4

# ─── Notifications ──────────────────────────────────────
aiosmtplib==3.0.1
python-telegram-bot==21.3

# ─── Utilities ──────────────────────────────────────────
python-dateutil==2.9.0
pytz==2024.1
loguru==0.7.2
tenacity==8.3.0
orjson==3.10.3

# ─── Monitoring ─────────────────────────────────────────
prometheus-fastapi-instrumentator==7.0.0

# ─── Dev / Testing ──────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.0
faker==25.2.0
